Improve R benchmark harness
Instead of calculating a fixed number of jobs, calculate one job per
core.  This avoids the issue of the last set of jobs being run on a
subset of the cores, if the number of jobs is not evenly divisible by
the number of cores.  The time for the run is divided by the number of
cores, thus giving a time/job throughput metric.

Due to changed benchmarks, introduce an updated reference.json file.

Add not to README.md how to calculate the geometric mean from the
results.